[Event Hubs] Fix endianness in partition key hashing
