+++ sdk/eventhub/event-hubs/src/impl/partitionAssigner.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { isDefined } from "../util/typeGuards";
-import { mapPartitionKeyToId } from "./patitionKeyToIdMapper";
+import { mapPartitionKeyToId } from "./partitionKeyToIdMapper";
 
 /**
  * @internal
