+++ sdk/eventhub/event-hubs/test/internal/partitionKeyHashMap.spec.ts
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { assert } from "chai";
+import { hashPartitionKey } from "../../src/impl/partitionKeyToIdMapper";
+
+/**
+ * These unit tests have been created from outputs received from the C# implementation
+ * of Jenkins lookup3 that the Event Hubs service uses.
+ */
+describe("hashPartitionKey", () => {
+  it("Generated hashes match the ones the service would have generated", () => {
+    const expectations: { Key: string; Hash: number }[] = require("./partitionKeyHashMap.json");
+    expectations.map(({ Key, Hash }) => assert.equal(hashPartitionKey(Key), Hash));
+  });
+});
